feat(db2): Add parser functions and generator methods
- Add parser functions: CHAR, DAYOFWEEK, DAYOFYEAR, POSSTR, VARCHAR_FORMAT
- Add generator transforms for StrPosition and TimeToStr
- Add extract_sql method for DAYOFWEEK/DAYOFYEAR handling
- Update to use separate generator file pattern from PR1
