+++ sqlglot/dialects/db2.py
@@ -2,8 +2,9 @@ from __future__ import annotations
 
 import typing as t
 
-from sqlglot import exp, generator, tokens
-from sqlglot.dialects.dialect import Dialect, NormalizationStrategy, rename_func
+from sqlglot import tokens
+from sqlglot.dialects.dialect import Dialect, NormalizationStrategy
+from sqlglot.generators.db2 import Db2 as Db2Generator
 from sqlglot.parsers.db2 import Db2Parser
 from sqlglot.tokens import TokenType
 
@@ -51,19 +52,4 @@ class Db2(Dialect):
         }
 
     Parser = Db2Parser
-
-    class Generator(generator.Generator):
-        TRANSFORMS = {
-            **generator.Generator.TRANSFORMS,
-            exp.StrPosition: rename_func("POSSTR"),
-            exp.TimeToStr: rename_func("VARCHAR_FORMAT"),
-        }
-
-        def extract_sql(self, expression: exp.Extract) -> str:
-            this = self.sql(expression, "this")
-            expression_sql = self.sql(expression, "expression")
-
-            if this.upper() in ("DAYOFWEEK", "DAYOFYEAR"):
-                return f"{this.upper()}({expression_sql})"
-
-            return f"EXTRACT({this} FROM {expression_sql})"
+    Generator = Db2Generator
